Scope tournament editing updates

+++ tabbycat/draw/consumers.py
@@ -7,9 +7,11 @@ from channels.layers import get_channel_layer
 
 from actionlog.models import ActionLogEntry
 from adjallocation.serializers import SimpleDebateAllocationSerializer, SimpleDebateImportanceSerializer
+from participants.models import Adjudicator, Team
 from tournaments.mixins import RoundWebsocketMixin
 from users.permissions import Permission
 from utils.mixins import SuperuserRequiredWebsocketMixin
+from venues.models import Venue
 from venues.serializers import SimpleDebateVenueSerializer
 
 from .models import Debate, DebateTeam
@@ -50,7 +52,8 @@ class BaseAdjudicatorContainerConsumer(SuperuserRequiredWebsocketMixin, RoundWeb
     def get_debates_or_panels(self, debates_or_panels):
         """ Retrieve either the debates or panels from the JSON id keys """
         ids = [id for (id, d_or_p) in debates_or_panels.items()]
-        debates_or_panels = list(self.model.objects.filter(id__in=ids))
+        debates_or_panels = list(self.model.objects.filter(
+            id__in=ids, round__tournament=self.tournament))
         # TODO: error handling if return items fewer/more than expected
         return debates_or_panels
 
@@ -85,6 +88,9 @@ class BaseAdjudicatorContainerConsumer(SuperuserRequiredWebsocketMixin, RoundWeb
             sent_allocation_ids = []
             for (position, position_ids) in sent_allocation.items():
                 sent_allocation_ids.extend(adj_id for adj_id in position_ids)
+            if Adjudicator.objects.filter(id__in=sent_allocation_ids,
+                    tournament=self.tournament).count() != len(set(sent_allocation_ids)):
+                continue
 
             # Delete adjudicators in the posted information
             self.delete_adjudicators(d_or_p, sent_allocation_ids)
@@ -155,6 +161,11 @@ class DebateEditConsumer(BaseAdjudicatorContainerConsumer):
         return super().receive_json(content)
 
     def modify_debate_teams(self, debate, sent_teams):
+        sent_team_ids = [team_id for team_id in sent_teams if team_id is not None]
+        if Team.objects.filter(id__in=sent_team_ids,
+                tournament=self.tournament).count() != len(set(sent_team_ids)):
+            return
+
         # Delete existing entries that won't be wanted (there shouldn't be any, but just in case)
         delete_count, deleted = debate.debateteam_set.exclude(side__in=self.tournament.sides).delete()
         logger.debug("Deleted %d debate teams from [%s]", deleted.get('draw.DebateTeam', 0), debate.matchup)
@@ -197,6 +208,11 @@ class DebateEditConsumer(BaseAdjudicatorContainerConsumer):
         changes = {int(c['id']): c for c in content[key]}
         debates = self.get_debates_or_panels(changes)
         for debate in debates:
+            if field_name == 'venue_id':
+                venue_id = changes[debate.id][content_name]
+                if venue_id is not None and not Venue.objects.filter(
+                        id=venue_id, tournament=self.tournament).exists():
+                    continue
             setattr(debate, field_name, changes[debate.id][content_name])
             debate.save()
